A temporary function for score calculation
This is a approximate score function, that calculates score for any round, the function is not 100% accurate, it shows some mistakes, but the distribution of the function is perfect, after testing lots of formulae, i made it get this distribution and a mean of 0.2-0.3 in preflop round and 0.3-0.4 in flop round and 0.5 in the next rounds. Which is what im exactly looking for. This maybe helpful.